# res/openapi/huki-ebook-openapi.generated.json
{
  "components": {
    "schemas": {
...
      "identity_ChangePasswordDto": {
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "example": "NewPassword123!",
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "type": "object"
      },
      "identity_ForgotPasswordDto": {
        "properties": {
          "email": {
            "example": "user@huki.com",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "type": "object"
      },
      "identity_LoginDto": {
        "properties": {
          "email": {
            "example": "user@huki.com",
            "type": "string"
          },
          "password": {
            "example": "Password123!",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
      "identity_RefreshTokenDto": {
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "type": "object"
      },
      "identity_RegisterDto": {
        "properties": {
          "email": {
            "example": "user@huki.com",
            "type": "string"
          },
          "fullName": {
            "example": "Nguyen Van A",
            "type": "string"
          },
          "password": {
            "description": "Min 8 chars, 1 uppercase, 1 number, 1 special char (@$!%*?&)",
            "example": "Password123!",
            "type": "string"
          },
          "phone": {
            "example": "0912345678",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "fullName"
        ],
        "type": "object"
      },
      "identity_ResetPasswordDto": {
        "properties": {
          "newPassword": {
            "example": "NewPassword123!",
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "newPassword"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "commerce_bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "community_bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "identity_bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "JWT-auth": {
        "bearerFormat": "JWT",
        "description": "Enter JWT token",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "type": "http"
      },
      "JWT-refresh": {
        "bearerFormat": "JWT",
        "description": "Enter JWT token",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "type": "http"
      },
      "shipping_bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "https://huki.ebook.com",
      "name": "HUKI Support",
      "url": "support@huki.ebook.com"
    },
    "description": "\n# HUKI EBOOK Platform API\n\nWelcome to the HUKI EBOOK API documentation. This API provides endpoints for managing books, orders, payments, and more.\n\n## Base URL\n`http://localhost:3000/api/v1`\n\n## Authentication\nMost endpoints require JWT Bearer token authentication.\n1. Register/Login to get access token\n2. Include header: `Authorization: Bearer <token>`\n\n## Services\n| Service | Port | Description |\n|---------|------|-------------|\n| Identity | 3001 | Auth, Users, Sessions |\n| Business | 3002 | Business, Stores, Members |\n| Commerce | 3003 | Books, Cart, Orders, Payments |\n| Shipping | 3004 | Shipments, Addresses |\n| Community | 3005 | Forum, Chat, Reviews |\n| Promotion | 3007 | Vouchers, Banners |\n\n## Common Headers\n| Header | Required | Description |\n|--------|----------|-------------|\n| Content-Type | Yes | application/json |\n| Authorization | No | Bearer token |\n| Idempotency-Key | For checkout | Unique key for idempotent operations |\n\n## Error Codes\n- 400: Bad Request\n- 401: Unauthorized\n- 403: Forbidden\n- 404: Not Found\n- 409: Conflict\n- 500: Server Error\n\n---\n*Last updated: 2026-08-24*\n    ",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://huki.ebook.com/terms",
...
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/change-password": {
      "patch": {
        "description": "Changes the user password after verifying the current password.",
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/identity_ChangePasswordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "400": {
            "description": "Current password incorrect"
          },
          "401": {
            "description": "Invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change password while logged in",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "description": "Sends a password reset email to the user if the email exists.",
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/identity_ForgotPasswordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Reset email sent if account exists"
          }
        },
        "summary": "Request password reset",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Authenticates user with email and password. Requires verified email.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/identity_LoginDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Invalid credentials or unverified email"
          }
        },
        "summary": "Login with credentials",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Revokes the refresh token and logs out from current device.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logout successful"
          },
          "401": {
            "description": "Invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout from current device",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "description": "Revokes all refresh tokens and logs out from all devices.",
        "operationId": "AuthController_logoutAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logged out from all devices"
          },
          "401": {
            "description": "Invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout from all devices",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "Returns the profile of the currently authenticated user.",
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user profile"
          },
          "401": {
            "description": "Invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Issues a new access token using a valid refresh token.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/identity_RefreshTokenDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Token refreshed"
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        },
        "summary": "Refresh access token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "Creates a new user account. Requires email verification.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/identity_RegisterDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Registration successful"
          },
          "400": {
            "description": "Invalid input data"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/auth/reset-password": {
      "post": {
        "description": "Resets the user password using a valid reset token.",
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/identity_ResetPasswordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Password reset successful"
          },
          "400": {
            "description": "Invalid or expired reset token"
          }
        },
        "summary": "Reset password with token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "description": "Verifies user email using token from verification email.",
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Email verified successfully"
          },
          "400": {
            "description": "Invalid or expired token"
          }
        },
        "summary": "Verify email address",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "servers": [
...
}